+++ src/Operations.jl
@@ -6,8 +6,6 @@ using UUIDs
 using Random: randstring
 import LibGit2, Dates, TOML
 
-import REPL
-using REPL.TerminalMenus
 using ..Types, ..Resolve, ..PlatformEngines, ..GitTools, ..MiniProgressBars
 import ..depots, ..depots1, ..devdir, ..set_readonly, ..Types.PackageEntry
 import ..Artifacts: ensure_artifact_installed, artifact_names, extract_all_hashes,
@@ -1376,7 +1374,8 @@ function _resolve(io::IO, env::EnvCache, registries::Vector{Registry.RegistryIns
 end
 
 function add(ctx::Context, pkgs::Vector{PackageSpec}, new_git=Set{UUID}();
-             preserve::PreserveLevel=default_preserve(), platform::AbstractPlatform=HostPlatform())
+             preserve::PreserveLevel=default_preserve(), platform::AbstractPlatform=HostPlatform(),
+             target::Symbol=:deps)
     assert_can_add(ctx, pkgs)
     # load manifest data
     for (i, pkg) in pairs(pkgs)
@@ -1384,20 +1383,55 @@ function add(ctx::Context, pkgs::Vector{PackageSpec}, new_git=Set{UUID}();
         is_dep = any(uuid -> uuid == pkg.uuid, [uuid for (name, uuid) in ctx.env.project.deps])
         pkgs[i] = update_package_add(ctx, pkg, entry, is_dep)
     end
-    foreach(pkg -> ctx.env.project.deps[pkg.name] = pkg.uuid, pkgs) # update set of deps
-    # resolve
-    pkgs, deps_map = _resolve(ctx.io, ctx.env, ctx.registries, pkgs, preserve, ctx.julia_version)
-    update_manifest!(ctx.env, pkgs, deps_map, ctx.julia_version)
-    new_apply = download_source(ctx)
-    fixup_ext!(ctx.env, pkgs)
 
-    # After downloading resolutionary packages, search for (Julia)Artifacts.toml files
-    # and ensure they are all downloaded and unpacked as well:
-    download_artifacts(ctx.env, platform=platform, julia_version=ctx.julia_version, io=ctx.io)
+    names = (p.name for p in pkgs)
+    target_field = if target == :deps
+        ctx.env.project.deps
+    elseif target == :weakdeps
+        ctx.env.project.weakdeps
+    elseif target == :extras
+        ctx.env.project.extras
+    else
+        pkgerror("Unrecognized target $(target)")
+    end
 
-    write_env(ctx.env) # write env before building
-    show_update(ctx.env, ctx.registries; io=ctx.io)
-    build_versions(ctx, union(new_apply, new_git))
+    foreach(pkg -> target_field[pkg.name] = pkg.uuid, pkgs) # update set of deps/weakdeps/extras
+
+    if target == :deps # nothing to resolve/install if it's weak or extras
+        # resolve
+        man_pkgs, deps_map = _resolve(ctx.io, ctx.env, ctx.registries, pkgs, preserve, ctx.julia_version)
+        update_manifest!(ctx.env, man_pkgs, deps_map, ctx.julia_version)
+        new_apply = download_source(ctx)
+        fixup_ext!(ctx.env, man_pkgs)
+
+        # After downloading resolutionary packages, search for (Julia)Artifacts.toml files
+        # and ensure they are all downloaded and unpacked as well:
+        download_artifacts(ctx.env, platform=platform, julia_version=ctx.julia_version, io=ctx.io)
+
+        # if env is a package add compat entries
+        if ctx.env.project.name !== nothing && ctx.env.project.uuid !== nothing
+            compat_names = String[]
+            for pkg in pkgs
+                haskey(ctx.env.project.compat, pkg.name) && continue
+                pkgversion = Base.thispatch(ctx.env.manifest[pkg.uuid].version)
+                set_compat(ctx.env.project, pkg.name, string(pkgversion))
+                push!(compat_names, pkg.name)
+            end
+            printpkgstyle(ctx.io, :Compat, """entries added for $(join(compat_names, ", "))""")
+        end
+        record_project_hash(ctx.env) # compat entries changed the hash after it was last recorded in update_manifest!
+
+        write_env(ctx.env) # write env before building
+        show_update(ctx.env, ctx.registries; io=ctx.io)
+        build_versions(ctx, union(new_apply, new_git))
+        Pkg._auto_precompile(ctx)
+    else
+        record_project_hash(ctx.env)
+        write_env(ctx.env)
+        names_str = join(names, ", ")
+        printpkgstyle(ctx.io, :Added, "$names_str to [$(target)]")
+    end
+    return
 end
 
 # Input: name, uuid, and path
@@ -1686,7 +1720,6 @@ function gen_subprocess_cmd(code::String, source_path::String; coverage, julia_a
         $(Base.julia_cmd())
         --code-coverage=$(coverage_arg)
         --color=$(Base.have_color === nothing ? "auto" : Base.have_color ? "yes" : "no")
-        --compiled-modules=$(Bool(Base.JLOptions().use_compiled_modules) ? "yes" : "no")
         --check-bounds=yes
         --warn-overwrite=yes
         --depwarn=$(Base.JLOptions().depwarn == 2 ? "error" : "yes")