+++ test/extensions.jl
@@ -54,7 +54,7 @@ using Test
     isolate(loaded_depot=false) do
         depot = mktempdir(); empty!(DEPOT_PATH); push!(DEPOT_PATH, depot)
         Pkg.activate(; temp=true)
-        Pkg.Registry.add(Pkg.RegistrySpec(path=joinpath(@__DIR__, "test_packages", "ExtensionExamples", "ExtensionRegistry")))
+        Pkg.Registry.add(path=joinpath(@__DIR__, "test_packages", "ExtensionExamples", "ExtensionRegistry"))
         Pkg.Registry.add("General")
         Pkg.add("HasExtensions")
         Pkg.test("HasExtensions", julia_args=`--depwarn=no`) # OffsetArrays errors from depwarn
@@ -66,7 +66,7 @@ using Test
         withenv("JULIA_PKG_PRECOMPILE_AUTO" => 0) do
             depot = mktempdir(); empty!(DEPOT_PATH); push!(DEPOT_PATH, depot)
             Pkg.activate(; temp=true)
-            Pkg.Registry.add(Pkg.RegistrySpec(path=joinpath(@__DIR__, "test_packages", "ExtensionExamples", "ExtensionRegistry")))
+            Pkg.Registry.add(path=joinpath(@__DIR__, "test_packages", "ExtensionExamples", "ExtensionRegistry"))
             Pkg.Registry.add("General")
             Pkg.add("HasDepWithExtensions")
         end
@@ -77,5 +77,20 @@ using Test
         @test occursin("OffsetArraysExt", out)
         @test occursin("HasExtensions", out)
         @test occursin("HasDepWithExtensions", out)
+    end
+    isolate(loaded_depot=false) do
+        withenv("JULIA_PKG_PRECOMPILE_AUTO" => 0) do
+            Pkg.activate(; temp=true)
+            Pkg.add("Example", target=:weakdeps)
+            proj = Pkg.Types.Context().env.project
+            @test isempty(proj.deps)
+            @test proj.weakdeps == Dict{String, Base.UUID}("Example" => Base.UUID("7876af07-990d-54b4-ab0e-23690620f79a"))
+
+            Pkg.activate(; temp=true)
+            Pkg.add("Example", target=:extras)
+            proj = Pkg.Types.Context().env.project
+            @test isempty(proj.deps)
+            @test proj.extras == Dict{String, Base.UUID}("Example" => Base.UUID("7876af07-990d-54b4-ab0e-23690620f79a"))
+        end
     end
 end